// frontend_restaurants/web/src/lib/types.ts
export type orderItems = orderItem[];

type OrderType = {
	orderId: number;
	type: string;
	orderTime: Date;
	orderItems: orderItems;
	noteFromUser: string;
	totalPrice: number;
	finishTime: Date;
	completeTime: Date;
	rating: number;
	comment: string;
};

export type IncomingOrderType = Omit<
	OrderType,
	'finishTime' | 'completeTime' | 'rating' | 'comment'
>;
// ...
export type IncomingOrderList = IncomingOrderType[];